removing FilterModel class, as React doesn't really support  class components
